Revert "update v2 lyrics"
